Fix package versioning
The `setuptools_scm` package is used for instantiating the package version at
build time. This requirement stems from the toplevel `pyproject.toml` file that
autospec doesn't parse yet... that file is the entry point for PEP 517 style
builds, which autospec will eventually need to support as more packages switch
to it.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>